+++ bukkit/src/main/java/de/maxhenkel/voicechat/compatibility/Compatibility1_12.java
@@ -5,7 +5,6 @@ import de.maxhenkel.voicechat.BukkitVersion;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.serializer.gson.GsonComponentSerializer;
 import net.kyori.adventure.key.Key;
-import org.bukkit.Server;
 import org.bukkit.entity.Player;
 
 public class Compatibility1_12 extends BaseCompatibility {
@@ -17,13 +16,6 @@ public class Compatibility1_12 extends BaseCompatibility {
     public static final BukkitVersion VERSION_1_12 = BukkitVersion.parseBukkitVersion("1.12-R0.1");
 
     public static final Compatibility1_12 INSTANCE = new Compatibility1_12();
-
-    @Override
-    public String getServerIp(Server server) throws Exception {
-        Object minecraftServer = callMethod(server, "getServer");
-        Class<?> minecraftServerClass = getBukkitClass("MinecraftServer");
-        return callMethod(minecraftServerClass, minecraftServer, "getServerIp");
-    }
 
     @Override
     public Key createNamespacedKey(String key) {