+++ bukkit/src/main/java/de/maxhenkel/voicechat/compatibility/Compatibility1_17.java
@@ -4,7 +4,6 @@ import com.mojang.brigadier.arguments.ArgumentType;
 import de.maxhenkel.voicechat.BukkitVersion;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.serializer.gson.GsonComponentSerializer;
-import org.bukkit.Server;
 import org.bukkit.entity.Player;
 
 import java.util.UUID;
@@ -15,12 +14,6 @@ public class Compatibility1_17 extends BaseCompatibility {
     public static final BukkitVersion VERSION_1_17 = BukkitVersion.parseBukkitVersion("1.17-R0.1");
 
     public static final Compatibility1_17 INSTANCE = new Compatibility1_17();
-
-    @Override
-    public String getServerIp(Server server) throws Exception {
-        Object dedicatedServerProperties = callMethod(server, "getProperties");
-        return getField(dedicatedServerProperties, "c");
-    }
 
     @Override
     public void sendMessage(Player player, Component component) {