+++ bukkit/changelog.md
@@ -1,1 +1,1 @@
-- Updated to 1.21.4
+- Improved server-ip detection